requirements.txt
streamlit
python-dotenv
langchain-core	
langchain-community
langchain-groq
langchain-google-genai	